fix: load first valid SDF record for bulk-docking pose overlays
Share _first_valid_mol_from_sdf across pose rehydrate and Ligand reload
so multi-record mock pose files work in get_poses() and protein.show().
